Export serial connection to another file
